applications/backend/services/server/src/pyproject.toml
[tool.poetry]
name = "Server"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
codefly-sdk = "^0.0.7"
psycopg2 = "^2.9.9"
redis = "^5.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
